% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{addSourceUrlVersion}
\alias{addSourceUrlVersion}
\title{Adds the SourceUrl and SourceVersion fields}
\usage{
addSourceUrlVersion(x)
}
\arguments{
\item{x}{A list of metadata fields, containing a FullUrl field.}
}
\value{
Update list of metadat fields
}
\description{
Adds the SourceUrl and SourceVersion fields
}

