% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{checkMetaDataList}
\alias{checkMetaDataList}
\title{Validity of an AH metadata list}
\usage{
checkMetaDataList(x, n)
}
\arguments{
\item{x}{A list of metadata fields.}

\item{n}{The expected number of AH data objects to be documented.}
}
\value{
\code{NULL}, if tested statements are
    \code{TRUE}. Otherwise throws an error.
}
\description{
Validity of an AH metadata list
}

