% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{fixMetaDataList}
\alias{fixMetaDataList}
\title{Fix length of metadata fields}
\usage{
fixMetaDataList(x, n)
}
\arguments{
\item{x}{A list of metadata fields}

\item{n}{The expected number of AH data objects to be documented.}
}
\value{
The updated and valid list of metadata fields.
}
\description{
Fix length of metadata fields
}

