% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{identicalRemLoc}
\alias{identicalRemLoc}
\title{Are the remote and local instances identical}
\usage{
identicalRemLoc(rem, loc)
}
\arguments{
\item{rem}{An instance of class \code{AnnotationHub}}

\item{loc}{An instance of class \code{AnnotationHubMetadata}}
}
\value{
A \code{logical}
}
\description{
Are the remote and local instances identical
}

