% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{makeAnnotationHubMetadata}
\alias{makeAnnotationHubMetadata}
\title{Create a list of AH resources of same type}
\usage{
makeAnnotationHubMetadata(x, resource = c("FASTA", "mzTab", "mzid", "mzML"))
}
\arguments{
\item{x}{A valid list of metadata fields.}

\item{resource}{What type of resource}
}
\value{
A list of AnnotationHubMetadata of same type
}
\description{
Create a list of AH resources of same type
}

