% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{orderMetaDataList}
\alias{orderMetaDataList}
\title{Order AH metadata by \code{RDataClass}}
\usage{
orderMetaDataList(x)
}
\arguments{
\item{x}{A list of metadata fields}
}
\value{
Ordered list of metadata fields.
}
\description{
Order AH metadata by \code{RDataClass}
}

