% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAHD.R
\docType{class}
\name{PAHD-class}
\alias{.PAHD}
\alias{PAHD-class}
\title{An S4 class for ProteomicsAnnotationHubData objects}
\description{
This class is a simple temporary container that extends the
\code{AnnotationHubMetadata}. Please read that documentation for
details. This class is likely to evolve in the future. See
\code{\link{PAHD}} for to construct these objects from dcf files.
}

