% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAHD.R
\name{PAHD}
\alias{PAHD}
\title{Prepare data for inclusion into AnnotationHub}
\usage{
PAHD(x, resourceDir)
}
\arguments{
\item{x}{One of mulitple ProteomicsAnnotationHubData dcf files.}

\item{resourceDir}{A \code{character} containing the directory
holding the files (for instance PRIDE). If missing, will be
inferred from the first \code{RDataPath} field of \code{x}.}
}
\value{
A \code{list} of \code{PAHD} objects that can be used to
    prepare and submit data to AnnotationHub. See
    \code{ProteomicsAnnotationHub()} for details.
}
\description{
Reads ProteomicsAnnotationHubData dcf files and prepares them for
inclusion into AnnotationHub. See
\code{ProteomicsAnnotationDataHub{}} for details and an example.
}
\details{
Note: Current limitation is that all the files are expected to
reside in a single resource directory.
}
\author{
Laurent Gatto
## example file for the PXD000001 data
f <- list.files(system.file("extdata", package = "ProteomicsAnnotationHubData"),
                full.names = TRUE)
PXD000001 <- PAHD(f)
length(PXD000001)
PXD000001[[1]]
}

