% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ProteomicsAnnotationHubData.R
\name{ProteomicsAnnotationHubData}
\alias{ProteomicsAnnotationHubData}
\title{Get started with ProteomicsAnnotationHubData}
\usage{
ProteomicsAnnotationHubData()
}
\value{
Used for its side-effect of open the package vignette.
}
\description{
Read the \code{ProteomicsAnnotationHubData} vignette to get
started with using Proteomics data from AnnotationHub and writing
new recipes.
}
\author{
Laurent Gatto
}

