% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAHD.R
\name{readPahdFiles}
\alias{readPahdFiles}
\title{Reads one or multiple PAHD template files}
\usage{
readPahdFiles(file)
}
\arguments{
\item{file}{A \code{character} with one of multiple file names.}
}
\value{
A \code{matrix} containing
}
\description{
Reads one or multiple ProteomicsAnnotationHubData dcf files into a
\code{matrix} that can be processed with
\code{\link{PAHD}}. Commnent lines starting with \code{#} will be
removed. See \code{ProteomicsAnnotationDataHub()} for details.
}
\examples{
## example file for the PXD000001 data
f <- list.files(system.file("extdata", package = "ProteomicsAnnotationHubData"),
                full.names = TRUE, pattern = "PXD000001.dcf")
readPahdFiles(f)
}
\author{
Laurent Gatto
}

