% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAHD.R
\name{writePahdTemplate}
\alias{writePahdTemplate}
\title{Write a ProteomicsAnnotationHubData template}
\usage{
writePahdTemplate(filename = "")
}
\arguments{
\item{filename}{The name of the file to write the template
in. Default is \code{""}, i.e. write output to the console.}
}
\value{
Use for its side effect of preparing a annotation
    template.
}
\description{
Writes a simple template in dcf format (like an R package
DESCRIPTION file) that, once completed (see
\code{ProteomicsAnnotationHub()} for details), can be imported
with \code{\link{readPahdFiles}} or directly processed with
\code{\link{PAHD}}. Note that these cdf files support comments (as
opposed to DESCRIPTION files). Lines starting with \code{#} will
be removed when parsed by \code{\link{readPahdFiles}}.
}
\examples{
writePahdTemplate()
}
\author{
Laurent Gatto <lg390@cam.ac.uk>
}

