% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ProteomicsAnnotationHubData.R
\docType{data}
\name{ProteomicsAnnotationHubData}
\alias{ProteomicsAnnotationHubData}
\alias{availableProteomicsAnnotationHubData}
\title{Get started with ProteomicsAnnotationHubData}
\format{\preformatted{ chr "PXD000001"
}}
\usage{
ProteomicsAnnotationHubData()

availableProteomicsAnnotationHubData
}
\value{
Used for its side-effect of opening the package
    vignette. A vector of experiment identifiers.
}
\description{
Read the \code{ProteomicsAnnotationHubData} vignette to get
started with using Proteomics data from AnnotationHub and writing
new recipes. Use \code{availableProteomicsAnnotationHubData()} to
get a vector of available experiments.
}
\examples{
availableProteomicsAnnotationHubData()
}
\author{
Laurent Gatto
}
\keyword{datasets}

