% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteomicsAnnotationHubData.R
\docType{data}
\name{ProteomicsAnnotationHubData}
\alias{ProteomicsAnnotationHubData}
\alias{availableProteomicsAnnotationHubData}
\alias{proteomicsAnnotationHubDataResources}
\title{Get started with ProteomicsAnnotationHubData}
\format{An object of class \code{character} of length 1.}
\usage{
ProteomicsAnnotationHubData()

availableProteomicsAnnotationHubData

proteomicsAnnotationHubDataResources
}
\value{
Used for its side-effect of opening the package
    vignette. A vector of experiment identifiers.
}
\description{
Read the \code{ProteomicsAnnotationHubData} vignette to get
started with using Proteomics data from AnnotationHub and writing
new recipes. Use \code{availableProteomicsAnnotationHubData()} to
get a vector of available experiments. Use
\code{proteomicsAnnotationHubDataResources()} to get a
vector of available resources.
}
\examples{
availableProteomicsAnnotationHubData
}
\author{
Laurent Gatto
}
\keyword{datasets}

